+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog of dask-geomodeling
 2.2.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Allow up to 1E-7 in the GeoTransform 'tilt' terms to account for possible
+  float32 imprecision.
 
 
 2.2.3 (2020-02-28)