+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog of dask-geomodeling
 2.2.4 (unreleased)
 ------------------
 
+- Allow up to 1E-7 in the GeoTransform 'tilt' terms to account for possible
+  float32 imprecision.
+
 - Handle Nones in geometry.field_operations.Classify and ClassifyFromColumns.
 
 - Validate if labels are unique in geometry.field_operations.Classify and