+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog of dask-geomodeling
 2.2.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added raster.spatial.Place.
 
 
 2.2.2 (2020-02-13)